requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
google-generativeai==0.8.3
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
matplotlib==3.9.2
numpy==2.1.1
gunicorn==22.0.0
Werkzeug==3.0.4
seaborn==0.13.2
sympy==1.13.3
bcrypt==4.1.3
supabase==2.5.0
